Quality of life: Add inttypes.h, fix header perms, va_copy and "_Bool"
PRId32 and friends are here. The rest of inttypes.h is not, but it is
a library issue and arguably not that important.

The headers are no longer executable. :P

__va_copy is aliased as va_copy.

enum bool has been renamed to _Bool and defines bool to it.
It isn't exactly _Bool, but it is close enough.